Switch to using off_t from size_t where 64-bit types matter.

+++ src/XrdOfs/XrdOfs.cc
@@ -906,7 +906,7 @@ int XrdOfsFile::write(XrdSfsAio *aiop)
 /*                               g e t M m a p                                */
 /******************************************************************************/
 
-int XrdOfsFile::getMmap(void **Addr, size_t &Size)         // Out
+int XrdOfsFile::getMmap(void **Addr, off_t &Size)         // Out
 /*
   Function: Return memory mapping for file, if any.
